fix: move Hero page back to pages namespace
